backend/requirements.txt
fastapi==0.104.1                 
uvicorn==0.24.0                 
pydantic==2.5.0                  
sentence-transformers==2.2.2     
torch==2.0.1                     
scikit-learn==1.3.0              
numpy==1.24.3                    
scipy==1.15.3                    
python-dotenv==1.0.0 